+++ framework/tests/value/RegexNonMatching.java
@@ -1,0 +1,7 @@
+// A test for the @DoesNotMatchRegex annotation.
+
+import org.checkerframework.common.value.qual.*;
+
+public class RegexNonMatching {
+  // TODO
+}