+++ docs/CHANGELOG.md
@@ -1,3 +1,15 @@
+Version 3.26.1 (November 1, 2022)
+---------------------------------
+
+**User-visible changes:**
+
+The Constant Value Checker supports new annotation @DoesNotMatchRegex.
+
+**Implementation details:**
+
+**Closed issues:**
+
+
 Version 3.26.0 (October 3, 2022)
 --------------------------------